refactor the viewport interaction module
